requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
idna==3.7
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pillow==10.4.0
requests==2.32.3
setuptools==70.3.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
